% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateC.R
\name{generateC}
\alias{generateC}
\title{generateC}
\usage{
generateC(windowSize, corrMatrix, outPath, outName)
}
\arguments{
\item{windowSize}{is the window size defined in terms of the number of SNPs}

\item{corrMatrix}{is obtained from the generateR function}

\item{outPath}{is a parameter that specifies the path of the result file}

\item{outName}{is a parameter that specifies the name of the result file}

\item{numThreads}{Number of parallel processes}
}
\value{
estimated correlation in the rotated space
}
\description{
generateC function will generate CovBand matrix
}
\examples{
   X <- as.matrix(read.table(SNPData))
   K <- Kinship(X, outPath, outName)
   VC <- varComp(Y, K, numThreads, outPath, outName)
   corrMatrix <- generateR(X, K, VC, outPath, outName)

   covBand <- generateC(windowSize, corrMatrix, outPath = "./", outName = "c.txt")
}
