% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateCovBand.R
\name{generateCovBand}
\alias{generateCovBand}
\title{generateCovBand function, R package version of MultiTrans}
\usage{
generateCovBand(windowSize, gR, outPath, outName)
}
\arguments{
\item{windowSize}{is the window size defined in terms of the number of SNPs.}

\item{gR}{is obtained from the generateR function}

\item{outPath}{is a parameter that specifies the path of the result file}

\item{outName}{is a parameter that specifies the name of the result file}
}
\value{
estimated correlation in the rotated space
}
\description{
generateCovBand function estimate correlation in the rotated space
}
\examples{
   X <- as.matrix(read.table(SNPData))
   K <- Kinship(X, outPath, outName)
   VC <- varComp(K, Y, X, outPath, outName)
   gR <- generateR(X, K, VC, outPath ="./", outName = "r.txt")

   covBand <- generateCovBand(1000, gR, outPath = "./", outName = "c.txt")
}
