% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateR.R
\name{generateR}
\alias{generateR}
\title{generateR function, R package version of MultiTrans}
\usage{
generateR(X, K, VC, outPath, outName)
}
\arguments{
\item{X}{is the SNP matrix, individual x snp}

\item{K}{is obtained from the Kinship function}

\item{VC}{is obtained from the varComp function}

\item{outPath}{is a parameter that specifies the path of the result file}

\item{outName}{is a parameter that specifies the name of the result file}
}
\value{
estimated correlation in the rotated space
}
\description{
generateR function estimate correlation in the rotated space
}
\examples{
   X <- as.matrix(read.table(SNPData))
   K <- Kinship(X, outPath, outName)
   VC <- varComp(Y, K, numThreads, outPath, outName)

   gR <- generateR(X, K, VC, outPath, outName)
}
