% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kinship.R
\name{Kinship}
\alias{Kinship}
\title{Kinship}
\usage{
Kinship(X, outPath, outName)
}
\arguments{
\item{X}{is the SNP matrix, individual x SNPs}

\item{outPath}{is a parameter that specifies the path of the result file}

\item{outName}{is a parameter that specifies the name of the result file}
}
\value{
kinship coefficient
}
\description{
Kinship function will estimate kinship coefficient
}
\examples{
   X <- as.matrix(read.table(SNPData))
   K <- Kinship(X, outPath, outName)
}
