% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Slide.R
\name{Slide}
\alias{Slide}
\title{generateCovBand function, R package version of MultiTrans}
\usage{
Slide(c, windowSize, maxstat, sampling, seed, sorted, threshold, outPath)
}
\arguments{
\item{c}{is obtained from the generateCovBand function}

\item{windowSize}{is the window size defined in terms of the number of SNPs.}

\item{maxstat}{max stat file}

\item{sampling}{100 used in MultiTrans paper, read MultiTrans paper for the detail, https://doi.org/10.1186/s13059-016-0903-6}

\item{seed}{123 used in MultiTrans paper, read MultiTrans paper for the detail, https://doi.org/10.1186/s13059-016-0903-6}

\item{sorted}{storing the sampled maximum statistics over the markers.}

\item{threshold}{is a text file containing pointwise p-values you want to correct, delimitered by space or newline.}

\item{outPath}{is a parameter that specifies the path of the result file}
}
\value{
estimated correlation in the rotated space
}
\description{
generateCovBand function estimate correlation in the rotated space
}
\examples{
   X <- as.matrix(read.table(SNPData))
   K <- Kinship(X, outPath, outName)
   VC <- varComp(K, Y, X, outPath, outName)
   gR <- generateR(X, K, VC, outPath ="./", outName = "r.txt")
   gC <- generateCovBand(1000, gR, outPath = "./", outName = "c.txt")

   slide <- Slide(gC, windowSize, "./maxstat", 100, 123, "./sorted", "./threshold.txt", "./result")
}
