% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varComp.R
\name{varComp}
\alias{varComp}
\title{varComp}
\usage{
varComp(K, Y, X, outPath, outName)
}
\arguments{
\item{K}{is obtained from the Kinship function}

\item{Y}{is the phenotype matrix, individual x phenotypes}

\item{X}{is the SNP matrix, individual x SNPs}

\item{outPath}{is a parameter that specifies the path of the result file}

\item{outName}{is a parameter that specifies the name of the result file}
}
\value{
Variance components Vg: genetic factor Ve: environment factor
}
\description{
varComp function will estimate variance components
}
\examples{
   X <- as.matrix(read.table(SNPData))
   Y <- as.matrix(read.table(GeneExpressionData))
   K <- Kinship(X, outPath, outName)

   VC <- varComp(K, Y, X, outPath, outName)
}
